+++ src/pages/admin/marketplace/edit.tsx
@@ -11,10 +11,11 @@ import Link from 'next/link'
 import Button, { ButtonSize, ButtonType } from '../../../components/Button'
 import { Marketplace } from './../../../types.d'
 import { useLogin } from '../../../hooks/login'
-import { initMarketplaceSDK } from './../../../modules/marketplace'
 import { AdminLayout } from 'src/layouts/Admin'
 import AdminMenu, { AdminMenuItemType } from '../../../components/AdminMenu'
 import { ReactElement } from 'react'
+import { MarketplaceClient } from '@holaplex/marketplace-js-sdk'
+import { Wallet } from '@metaplex/js'
 
 const SUBDOMAIN = process.env.MARKETPLACE_SUBDOMAIN
 
@@ -136,7 +137,7 @@ const AdminEditMarketplace = ({ marketplace }: AdminEditMarketplaceProps) => {
       return
     }
 
-    const client = initMarketplaceSDK(connection, wallet)
+    const client = new MarketplaceClient(connection, wallet as Wallet)
 
     toast('Saving changes...')
