+++ src/pages/nfts/[address].tsx
@@ -1,27 +1,9 @@
 import { gql, useQuery, QueryResult, OperationVariables } from '@apollo/client'
-import { AuctionHouseProgram } from '@metaplex-foundation/mpl-auction-house'
-import { MetadataProgram } from '@metaplex-foundation/mpl-token-metadata'
 import { useConnection, useWallet } from '@solana/wallet-adapter-react'
-import {
-  PublicKey,
-  SYSVAR_INSTRUCTIONS_PUBKEY,
-  Transaction,
-  TransactionInstruction,
-} from '@solana/web3.js'
 import { useRouter } from 'next/router'
 import { NextPage, NextPageContext } from 'next'
 import { AppProps } from 'next/app'
-import {
-  any,
-  concat,
-  intersection,
-  isEmpty,
-  isNil,
-  map,
-  or,
-  pipe,
-  prop,
-} from 'ramda'
+import { any, intersection, isEmpty, isNil, map, or, pipe, prop } from 'ramda'
 import { useForm } from 'react-hook-form'
 import Link from 'next/link'
 import { toast } from 'react-toastify'
@@ -31,17 +13,14 @@ import Button, { ButtonType } from '../../components/Button'
 import { useLogin } from '../../hooks/login'
 import { Listing, Marketplace, Nft, Offer, GetNftData } from '../../types.d'
 import { ReactElement } from 'react'
+import { ListingsClient } from '@holaplex/marketplace-js-sdk'
+import { Wallet } from '@metaplex/js'
+import {
+  Nft as NftFromSdk,
+  Listing as ListingFromSdk,
+} from '@holaplex/marketplace-js-sdk'
 
 const SUBDOMAIN = process.env.MARKETPLACE_SUBDOMAIN
-
-const {
-  createPublicBuyInstruction,
-  createExecuteSaleInstruction,
-  createCancelInstruction,
-  createPrintBidReceiptInstruction,
-  createCancelListingReceiptInstruction,
-  createPrintPurchaseReceiptInstruction,
-} = AuctionHouseProgram.instructions
 
 interface GetNftPage {
   marketplace: Marketplace | null
@@ -226,7 +205,8 @@ const NftShow: NextPage<NftPageProps> = ({
 }) => {
   const cancelListingForm = useForm()
   const buyNowForm = useForm()
-  const { publicKey, signTransaction } = useWallet()
+  const wallet = useWallet()
+  const { publicKey, signTransaction } = wallet
   const { connection } = useConnection()
   const login = useLogin()
 
@@ -240,193 +220,18 @@ const NftShow: NextPage<NftPageProps> = ({
       return
     }
 
-    const auctionHouse = new PublicKey(marketplace.auctionHouse.address)
-    const authority = new PublicKey(marketplace.auctionHouse.authority)
-    const auctionHouseFeeAccount = new PublicKey(
-      marketplace.auctionHouse.auctionHouseFeeAccount
-    )
-    const treasuryMint = new PublicKey(marketplace.auctionHouse.treasuryMint)
-    const seller = new PublicKey(listing.seller)
-    const tokenMint = new PublicKey(nft.mintAddress)
-    const auctionHouseTreasury = new PublicKey(
-      marketplace.auctionHouse.auctionHouseTreasury
-    )
-    const listingReceipt = new PublicKey(listing.address)
-    const sellerPaymentReceiptAccount = new PublicKey(listing.seller)
-    const sellerTradeState = new PublicKey(listing.tradeState)
-    const buyerPrice = listing.price.toNumber()
-    const tokenAccount = new PublicKey(nft.owner.associatedTokenAccountAddress)
-
-    const [metadata] = await MetadataProgram.findMetadataAccount(tokenMint)
-
-    const [escrowPaymentAccount, escrowPaymentBump] =
-      await AuctionHouseProgram.findEscrowPaymentAccountAddress(
-        auctionHouse,
-        publicKey
-      )
-
-    const [buyerTradeState, tradeStateBump] =
-      await AuctionHouseProgram.findPublicBidTradeStateAddress(
-        publicKey,
-        auctionHouse,
-        treasuryMint,
-        tokenMint,
-        buyerPrice,
-        1
-      )
-    const [freeTradeState, freeTradeStateBump] =
-      await AuctionHouseProgram.findTradeStateAddress(
-        seller,
-        auctionHouse,
-        tokenAccount,
-        treasuryMint,
-        tokenMint,
-        0,
-        1
-      )
-    const [programAsSigner, programAsSignerBump] =
-      await AuctionHouseProgram.findAuctionHouseProgramAsSignerAddress()
-    const [buyerReceiptTokenAccount] =
-      await AuctionHouseProgram.findAssociatedTokenAccountAddress(
-        tokenMint,
-        publicKey
-      )
-
-    const [bidReceipt, bidReceiptBump] =
-      await AuctionHouseProgram.findBidReceiptAddress(buyerTradeState)
-    const [purchaseReceipt, purchaseReceiptBump] =
-      await AuctionHouseProgram.findPurchaseReceiptAddress(
-        sellerTradeState,
-        buyerTradeState
-      )
-
-    const publicBuyInstructionAccounts = {
-      wallet: publicKey,
-      paymentAccount: publicKey,
-      transferAuthority: publicKey,
-      treasuryMint,
-      tokenAccount,
-      metadata,
-      escrowPaymentAccount,
-      authority,
-      auctionHouse,
-      auctionHouseFeeAccount,
-      buyerTradeState,
-    }
-    const publicBuyInstructionArgs = {
-      tradeStateBump,
-      escrowPaymentBump,
-      buyerPrice,
-      tokenSize: 1,
-    }
-
-    const executeSaleInstructionAccounts = {
-      buyer: publicKey,
-      seller,
-      tokenAccount,
-      tokenMint,
-      metadata,
-      treasuryMint,
-      escrowPaymentAccount,
-      sellerPaymentReceiptAccount,
-      buyerReceiptTokenAccount,
-      authority,
-      auctionHouse,
-      auctionHouseFeeAccount,
-      auctionHouseTreasury,
-      buyerTradeState,
-      sellerTradeState,
-      freeTradeState,
-      programAsSigner,
-    }
-
-    const executeSaleInstructionArgs = {
-      escrowPaymentBump,
-      freeTradeStateBump,
-      programAsSignerBump,
-      buyerPrice,
-      tokenSize: 1,
-    }
-
-    const printBidReceiptAccounts = {
-      bookkeeper: publicKey,
-      receipt: bidReceipt,
-      instruction: SYSVAR_INSTRUCTIONS_PUBKEY,
-    }
-    const printBidReceiptArgs = {
-      receiptBump: bidReceiptBump,
-    }
-
-    const printPurchaseReceiptAccounts = {
-      bookkeeper: publicKey,
-      purchaseReceipt,
-      bidReceipt,
-      listingReceipt,
-      instruction: SYSVAR_INSTRUCTIONS_PUBKEY,
-    }
-    const printPurchaseReceiptArgs = {
-      purchaseReceiptBump,
-    }
-
-    const publicBuyInstruction = createPublicBuyInstruction(
-      publicBuyInstructionAccounts,
-      publicBuyInstructionArgs
-    )
-    const printBidReceiptInstruction = createPrintBidReceiptInstruction(
-      printBidReceiptAccounts,
-      printBidReceiptArgs
-    )
-    const executeSaleInstruction = createExecuteSaleInstruction(
-      executeSaleInstructionAccounts,
-      executeSaleInstructionArgs
-    )
-    const printPurchaseReceiptInstruction =
-      createPrintPurchaseReceiptInstruction(
-        printPurchaseReceiptAccounts,
-        printPurchaseReceiptArgs
-      )
-
-    const txt = new Transaction()
-
-    txt
-      .add(publicBuyInstruction)
-      .add(printBidReceiptInstruction)
-      .add(
-        new TransactionInstruction({
-          programId: AuctionHouseProgram.PUBKEY,
-          data: executeSaleInstruction.data,
-          keys: concat(
-            executeSaleInstruction.keys,
-            nft.creators.map((creator) => ({
-              pubkey: new PublicKey(creator.address),
-              isSigner: false,
-              isWritable: true,
-            }))
-          ),
-        })
-      )
-      .add(printPurchaseReceiptInstruction)
-
-    txt.recentBlockhash = (await connection.getRecentBlockhash()).blockhash
-    txt.feePayer = publicKey
-
-    let signed: Transaction | undefined = undefined
-
-    try {
-      signed = await signTransaction(txt)
-    } catch (e: any) {
-      toast.error(e.message)
-      return
-    }
-
-    let signature: string | undefined = undefined
-
     try {
       toast('Sending the transaction to Solana.')
 
-      signature = await connection.sendRawTransaction(signed.serialize())
-
-      await connection.confirmTransaction(signature, 'confirmed')
+      const listingsClient = new ListingsClient(
+        connection,
+        wallet as Wallet,
+        marketplace.auctionHouse
+      )
+      await listingsClient.buy({
+        listing: listing as ListingFromSdk,
+        nft: nft as NftFromSdk,
+      })
 
       toast.success('The transaction was confirmed.')
 
@@ -446,78 +251,18 @@ const NftShow: NextPage<NftPageProps> = ({
       return
     }
 
-    const auctionHouse = new PublicKey(marketplace.auctionHouse.address)
-    const authority = new PublicKey(marketplace.auctionHouse.authority)
-    const auctionHouseFeeAccount = new PublicKey(
-      marketplace.auctionHouse.auctionHouseFeeAccount
-    )
-    const tokenMint = new PublicKey(nft.mintAddress)
-    const treasuryMint = new PublicKey(marketplace.auctionHouse.treasuryMint)
-    const receipt = new PublicKey(listing.address)
-    const tokenAccount = new PublicKey(nft.owner.associatedTokenAccountAddress)
-
-    const buyerPrice = listing.price.toNumber()
-
-    const [tradeState] = await AuctionHouseProgram.findTradeStateAddress(
-      publicKey,
-      auctionHouse,
-      tokenAccount,
-      treasuryMint,
-      tokenMint,
-      buyerPrice,
-      1
-    )
-
-    const cancelInstructionAccounts = {
-      wallet: publicKey,
-      tokenAccount,
-      tokenMint,
-      authority,
-      auctionHouse,
-      auctionHouseFeeAccount,
-      tradeState,
-    }
-    const cancelInstructionArgs = {
-      buyerPrice,
-      tokenSize: 1,
-    }
-
-    const cancelListingReceiptAccounts = {
-      receipt,
-      instruction: SYSVAR_INSTRUCTIONS_PUBKEY,
-    }
-
-    const cancelInstruction = createCancelInstruction(
-      cancelInstructionAccounts,
-      cancelInstructionArgs
-    )
-    const cancelListingReceiptInstruction =
-      createCancelListingReceiptInstruction(cancelListingReceiptAccounts)
-
-    const txt = new Transaction()
-
-    txt.add(cancelInstruction).add(cancelListingReceiptInstruction)
-
-    txt.recentBlockhash = (await connection.getRecentBlockhash()).blockhash
-    txt.feePayer = publicKey
-
-    let signed: Transaction | undefined = undefined
-
-    try {
-      signed = await signTransaction(txt)
-    } catch (e: any) {
-      toast.error(e.message)
-      return
-    }
-
-    let signature: string | undefined = undefined
-
     try {
       toast('Sending the transaction to Solana.')
 
-      signature = await connection.sendRawTransaction(signed.serialize())
-
-      await connection.confirmTransaction(signature, 'confirmed')
+      const listingsClient = new ListingsClient(
+        connection,
+        wallet as Wallet,
+        marketplace.auctionHouse
+      )
+      await listingsClient.cancel({
+        listing: listing as ListingFromSdk,
+        nft: nft as NftFromSdk,
+      })
 
       toast.success('The transaction was confirmed.')
